+++ src/main/java/br/com/escola/submissao/interfaces/dto/AvaliarSubmissaoRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record AvaliarSubmissaoRequest(
-        @NotNull Double nota,
+        @NotNull Long nota,
         @NotBlank String comentarioProfessor
 ) {
 }
